requirements.txt
Django
uims-api
djangorestframework
Pillow
django-cors-headers

#heroku
gunicorn
whitenoise
